move derp.go to derp module
